Update kookoo tests

+++ corehq/apps/ivr/tests/util.py
@@ -1,8 +1,11 @@
-from casexml.apps.case.models import CommCareCase
 from corehq.apps.ivr.models import Call
 from corehq.apps.sms.mixin import VerifiedNumber
 from corehq.apps.sms.models import INCOMING
 from corehq.apps.sms.util import register_sms_contact
+from corehq.form_processor.backends.sql.dbaccessors import CaseAccessorSQL
+from corehq.form_processor.interfaces.dbaccessors import CaseAccessors
+from corehq.form_processor.tests.utils import run_with_all_backends
+from corehq.form_processor.utils.general import should_use_sql_backend
 from django.test import TestCase
 
 
@@ -26,14 +29,16 @@ class LogCallTestCase(TestCase):
     def setUp(self):
         self.domain = 'test-log-call-domain'
         self.delete_call_logs(self.domain)
-        self.case_id = register_sms_contact(self.domain, 'participant', 'test',
-            'system', self.phone_number)
-        self.case = CommCareCase.get(self.case_id)
 
     def tearDown(self):
         self.delete_call_logs(self.domain)
         VerifiedNumber.by_phone(self.phone_number).delete()
-        self.case.delete()
+
+    def delete_case(self):
+        if should_use_sql_backend(self.domain):
+            CaseAccessorSQL.hard_delete_cases(self.domain, [self.case.case_id])
+        else:
+            self.case.delete()
 
     def simulate_inbound_call(self, phone_number):
         """
@@ -48,7 +53,12 @@ class LogCallTestCase(TestCase):
         """
         raise NotImplementedError("Please implement this method")
 
+    @run_with_all_backends
     def test_log_call(self):
+        self.case_id = register_sms_contact(self.domain, 'participant', 'test',
+            'system', self.phone_number)
+        self.case = CaseAccessors(self.domain).get_case(self.case_id)
+
         if self.__class__ == LogCallTestCase:
             # The test runner picks up this base class too, but we only
             # want to run the test on subclasses.
@@ -61,5 +71,7 @@ class LogCallTestCase(TestCase):
 
         call = Call.by_domain(self.domain)[0]
         self.assertEqual(call.couch_recipient_doc_type, 'CommCareCase')
-        self.assertEqual(call.couch_recipient, self.case.get_id)
+        self.assertEqual(call.couch_recipient, self.case.case_id)
         self.assertEqual(call.direction, INCOMING)
+
+        self.delete_case()
